requirements.txt
pdfplumber==0.10.2
python-dotenv==1.0.0
openai==0.27.8
jinja2
pydantic